Rearrange tab close buttons
With the verticaltabs-box on the right, autohide enabled and the buttons
placed on the far-right, the close button of a tab will be right underneath
the pointer when moving it over the box to expand it.
This way there will probably be some frustrated users who close tabs on
accident.

I placed them between the Title and the Thumbnail canvas for now, since I
think thats a god compromiss between reverting the whole UI dependent on
showtabsright or putting the close button right at the start of the row,
where it somhow looks displaced.