Reverse UI on showtabsright=true
Instead of repositioning just the tab-close buttons, this will inverse the
whole UI when showtabsright is set to true.
This way there is no need of change in the UI on the left and this also
fixes some similar problem with the "New Tab" button when right-aligned.

Currently this implementation is still buggy, because when authide option
is enabled, the thumbnail canvas is moved to the left side again.
I couldn't figure out why exactly that happens.